Arrumei as pastas e criei o PopUP
e Tambem criei o folder do popup e o componete da scree popUp
