--- src/screens/Add.jsx
@@ -1,22 +1,0 @@
-import { StyleSheet, Text, View } from 'react-native';
-
-export default function Add() {
-  return (
-    <View style={styles.container}>
-      <Text style= {styles.title}>add kcal e agua</Text>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: 'blue',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  title:{
-    fontSize:24,
-    fontWeight: 'bold',
-  }
-});
