--- src/screens/Historico.tsx
@@ -1,22 +1,0 @@
-import { StyleSheet, Text, View } from 'react-native';
-
-export default function Historico() {
-  return (
-    <View style={styles.container}>
-      <Text style= {styles.title}>Historico</Text>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: 'green',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  title:{
-    fontSize:24,
-    fontWeight: 'bold',
-  }
-});
